+++ src/app/app.module.ts
@@ -152,6 +152,7 @@ import { TreeSampleComponent } from './tree/tree.sample';
 import { GridColumnTypesSampleComponent } from './grid-column-types/grid-column-types.sample';
 import { AccordionSampleComponent } from './accordion/accordion.sample';
 import { GridLocalizationSampleComponent } from './grid-localization/grid-localization.sample';
+import { TreeGridGroupBySampleComponent } from './tree-grid-groupby/tree-grid-groupby.sample';
 import { PaginationSampleComponent } from './pagination/pagination.component';
 
 const components = [
@@ -246,6 +247,7 @@ const components = [
     TreeGridFlatDataSampleComponent,
     TreeGridWithTransactionsComponent,
     TreeGridLoadOnDemandSampleComponent,
+    TreeGridGroupBySampleComponent,
     CustomContentComponent,
     ColorsSampleComponent,
     AnimationsSampleComponent,