+++ src/app/app.routing.ts
@@ -92,6 +92,7 @@ import { GridUpdatesComponent } from './grid-updates-test/grid-updates.component
 import { TreeSampleComponent } from './tree/tree.sample';
 import { AccordionSampleComponent } from './accordion/accordion.sample';
 import { GridLocalizationSampleComponent } from './grid-localization/grid-localization.sample';
+import { TreeGridGroupBySampleComponent } from './tree-grid-groupby/tree-grid-groupby.sample';
 import { PaginationSampleComponent } from './pagination/pagination.component';
 
 const appRoutes = [
@@ -372,6 +373,10 @@ const appRoutes = [
     {
         path: 'treeGridFlatData',
         component: TreeGridFlatDataSampleComponent
+    },
+    {
+        path: 'treeGridGroupBy',
+        component: TreeGridGroupBySampleComponent
     },
     {
         path: 'treeGridLoadOnDemand',